Move detail::assert_if_is_constraint to MathematicalProgram
Fix #7020
